requirements.txt
bcrypt==4.0.1
cryptography
fastapi
psycopg2-binary
PyMySQL
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib
pytest
httpx
pytest-asyncio
aiofiles
jinja2

langchain
langchain_core
langchain_groq
langchain_google_genai
langchain_huggingface
python-dotenv
scikit-learn
pandas 
numpy
matplotlib